HBASE-13847 getWriteRequestCount function in HRegionServer uses int variable to return the count (Abhilash)

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/regionserver/HRegionServer.java
@@ -1114,7 +1114,7 @@ public class HRegionServer extends HasThread implements
    * @return Current write count for all online regions.
    */
   private long getWriteRequestCount() {
-    int writeCount = 0;
+    long writeCount = 0;
     for (Map.Entry<String, Region> e: this.onlineRegions.entrySet()) {
       writeCount += e.getValue().getWriteRequestsCount();
     }
